[refac] App.tsx 절대경로 수정 및 쓸데없는 주석 삭제
