+++ src/page/LogoutButton.tsx
@@ -1,9 +1,9 @@
 import React, { type ReactNode } from "react";
 import styled from "styled-components";
 import {useNavigate} from 'react-router-dom';
-import * as ROUTES from "../../constants/routes";
-import {logout} from "../../api";
-import {removeToken} from "../../util/AuthFunctions";
+import * as ROUTES from "../lib/routes";
+import {logout} from "../api";
+import {removeToken} from "../util/AuthFunctions";
 
 const Button = styled.button`
   position: fixed;
