+++ src/page/attendance/student/index.ts
@@ -1,0 +1,2 @@
+export * from './AttendanceModal';
+export * from './OpenLecturesPage';
