+++ src/page/enrollment/EnrollmentModal.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect} from "react";
 import styled from "styled-components";
 import {AnimatePresence, motion} from "framer-motion";
-import {requestEnrollment} from "../../../api";
+import {requestEnrollment} from "../../api";
 
 const Overlay = styled(motion.div)`
   position: fixed;
