+++ src/page/enrollment/index.ts
@@ -1,0 +1,2 @@
+export * from './EnrollmentModal';
+export * from './EnrollmentPage';
