+++ src/page/enrollment/manageEnrollment/EnrollmentManagePage.tsx
@@ -1,10 +1,12 @@
 import React, {useState, useEffect} from 'react';
 import styled from "styled-components";
-import {getStudentsEnrolledLectures} from "../../../../api";
-import LoadingSpinner from "../../../spinner/LoadingSpinner";
-import LectureRow from "../LectureRow";
-import * as ROUTES from "../../../../constants/routes";
+import {getLectureEnrollmentInfo, getLecturersOwnedLectures} from "../../../api";
+import LoadingSpinner from "../../../component/spinner/LoadingSpinner";
+import EnrollmentManageModal from "./EnrollmentManageModal";
+import LectureRow from "../../lecture/LectureRow";
+import * as ROUTES from "../../../lib/routes";
 import {useNavigate} from "react-router-dom";
+import {checkUserHasRole} from "../../../util/AuthFunctions";
 
 const LecturesContainer = styled.div`
   min-width: 40vw;
@@ -59,13 +61,19 @@ const Title = styled.button`
   color: black;
 `
 
-export const StudentsLecturesPage = () => {
+export const EnrollmentManagePage = () => {
     const [lectures, setLectures] = useState<Lecture[] | null>(null);
-    const [loading, setLoading] = useState(false);
+    const [currentLecture, setCurrentLecture] = useState<Lecture | null>(null);
+    const [isModalOpen, setModalOpen] = useState(false);
     const navigate = useNavigate();
 
     useEffect(() => {
-        getStudentsEnrolledLectures()
+        if (!checkUserHasRole(['ROLE_ADMIN', 'ROLE_LECTURER'])) {
+            navigate(ROUTES.LOGIN);
+            return;
+        }
+
+        getLecturersOwnedLectures()
             .then(lectureList => {
                 if (lectureList) {
                     setLectures(lectureList);
@@ -73,27 +81,29 @@ export const StudentsLecturesPage = () => {
                     alert('에러 발생! 관리자에게 문의하세요');
                     navigate(ROUTES.MAIN_PAGE);
                 }
-                setLoading(true);
             })
     }, []);
 
     return (
-        loading ?
-        lectures && lectures.length !== 0 ?
-                <LecturesContainer>
-                    <Title> 내 강의 </Title>
-                    <LectureTable>
-                        {Object.values(lectures).map((lecture, index) => {
+        lectures ?
+            <LecturesContainer>
+                <Title>강사 개설 강의 목록</Title>
+                <EnrollmentManageModal isOpen={isModalOpen} close={setModalOpen}
+                                       lecture={currentLecture ? currentLecture : null}/>
+                <LectureTable>
+                    {
+                        Object.values(lectures).map((lecture, index) => {
                             return (
-                                <LectureRow key={lecture.lectureId} index={index} lecture={lecture}/>
+                                <LectureRow key={lecture.lectureId} index={index} lecture={lecture}
+                                    onClick={() => {
+                                        setCurrentLecture(lecture);
+                                        setModalOpen(true);
+                                    }}/>
                             )
-                        })}
-                    </LectureTable>
-                </LecturesContainer>
-                :
-                <LecturesContainer>
-                    <Title> 수강중인 강의가 없습니다. </Title>
-                </LecturesContainer>
+                        })
+                    }
+                </LectureTable>
+            </LecturesContainer>
             :
             <LoadingSpinner/>
     );
