+++ src/page/enrollment/manageEnrollment/StudentEnrollmentStateRow.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react';
 import styled from "styled-components";
-import {getLectureEnrollmentInfo} from "../../../../api";
-import * as ROUTES from "../../../../constants/routes";
+import {getLectureEnrollmentInfo} from "../../../api";
+import * as ROUTES from "../../../lib/routes";
 
 const Student = styled.div`
   display: grid;
