+++ src/page/index.ts
@@ -1,0 +1,8 @@
+export * from './AgreementModal';
+export * from './LoginPage';
+export * from './LogoutButton';
+export * from './MainPage';
+export * from './SignUpPage';
+export * from './lecture';
+export * from './enrollment';
+export * from './attendance';
