+++ src/page/lecture/LectureRow.tsx
