+++ src/page/lecture/LecturesPage.tsx
@@ -1,8 +1,8 @@
 import React, {useState, useEffect} from 'react';
-import LoadingSpinner from "../../spinner/LoadingSpinner";
+import LoadingSpinner from "../../component/spinner/LoadingSpinner";
 import {StudentsLecturesPage} from "./student/StudentsLecturesPage";
-import {checkAndGetUserRole} from "../../../util/AuthFunctions";
-import * as ROUTES from "../../../constants/routes";
+import {checkAndGetUserRole} from "../../util/AuthFunctions";
+import * as ROUTES from "../../lib/routes";
 import {useNavigate} from "react-router-dom";
 import {LecturerLecturesPage} from "./lecturer/LecturerLecturesPage";
