+++ src/page/lecture/index.ts
@@ -1,0 +1,2 @@
+export * from './LecturesPage';
+export * from './LectureRow';
