+++ src/page/lecture/lecturer/index.ts
@@ -1,0 +1,5 @@
+export * from './LectureCreatePage';
+export * from './LectureModalWithStudents';
+export * from './LecturerLectureRow';
+export * from './LecturerLecturesPage';
+export * from './StudentRow';
