+++ src/page/lecture/student/index.ts
@@ -1,0 +1,1 @@
+export * from './StudentsLecturesPage';
