+++ src/ui/layout/Borderbox.tsx
@@ -3,14 +3,19 @@ import { ComponentProps } from "react";
 import { PaddingSize } from "./Padding";
 import { Padding } from "./Padding";
 
-
-
 interface BorderBoxProps extends ComponentProps<'div'>{
     fullWidth: boolean;
     padding?: PaddingSize;
     shadow?: boolean;
     height?: number;
 }
+
+/**
+ * @param fullWidth 전체 너비를 사용할것인지
+ * @param padding 패딩값 number : 상하좌우 [number, number] : 상하, 좌우 [number, number, number, number]
+ * @param shadow 섀도우를 넣을것인지?
+ * @param heigth 높이 
+ */
 
 type Props = Partial<BorderBoxProps>;
